ci(quality): warnings-as-errors gate, sanitizers, smoke, Python lint/tests
- EVENGINE_ENABLE_STRICT_WARNINGS option: MSVC /W4 /WX, GCC/Clang
  -Wall -Wextra -Werror, with documented noise suppressions and
  -isystem/-external:I for third-party headers; enabled on the
  Windows/Linux debug CI jobs.
- Fix warnings across engine, modules and tests exposed by the gate,
  including real bugs (incomplete-type deletes, Timespan microsecond
  overflow, overload hiding, enum switch gaps, dead code).
- Add Linux ASan+UBSan job, example smoke step, python-tools test
  job, ruff config (scripts+tools), weekly Linux coverage job and a
  CodeQL workflow.

+++ src/engine/cmdline/DevServer.cpp
@@ -319,7 +319,7 @@ int Cmdline::DevServer(std::string path, int port) {
     }
 
     DevFileServer server;
-    if (!server.start(port, path.empty() ? std::string(".") : path)) {
+    if (!server.start(static_cast<uint16_t>(port), path.empty() ? std::string(".") : path)) {
         cerr << "eve dev: cannot bind port " << port << endl;
         return 2;
     }
